+++ langgraph.json
@@ -7,7 +7,7 @@
     "coder": "./src/coder/lg_server.py:graph",
     "tester": "./src/tester/graph.py:graph",
     "code_reviewer": "./src/code_reviewer/graph.py:graph",
-    "requirement_gatherer": "./src/requirement_gatherer/graph_v2.py:graph",
+    "requirement_gatherer": "./src/requirement_gatherer/graph.py:graph",
     "task_manager": "./src/task_manager/graph.py:graph"
   },
   "env": ".env",
