+++ src/requirement_gatherer/prompts.py
@@ -1,6 +1,5 @@
 """Define default prompts."""
 
-
 SYSTEM_PROMPT = """
 # Product‐Requirement Gatherer — **System Prompt**
 
@@ -181,42 +180,4 @@ If NO questions are pending AND the final detailed requirements summary report h
 {user_info}
 
 System Time: {time}
-"""
-
-EVALUATOR_SYSTEM_PROMPT = """
-# Requirement‑Gathering Evaluator — System Prompt (Compact)
-
-You judge whether **Product‑requirement_gatherer** has fully met its own rules.
-
----
-
-## 1. Goals
-1. **Completeness** – mandatory files present & filled.  
-2. **Accuracy** – no contradictions with user answers.  
-3. **Quality** – smart question flow, risks logged, proper Markdown.  
-
----
-
-## 2. You Receive
-* `conversation_history`  
-* `repo_state` (Markdown files)  
-* `project_type` (“hobby” | “full_product”)
-
----
-
-## 3. Return ONE YAML Block
-
-```
-verdict: "<Completed | needs_more >"
-```
---- 
-
-### 4. Pass Criteria (Quick)
-
-- **Hobby** → `productOverview.md`, `functionalReqs.md`
-- **Full Product** → all Core Files from the Gatherer's spec
-- Each file has **at least 2 meaningful lines**
-- **Risks** include both an **owner** and a **due date**
-- Markdown formatting and `upsert_memory` usage are correct
-- You **only judge** — do **not** ask the user more questions
 """
